requirements.txt
flask
telebot
python-dotenv
psycopg2-binary
rapidfuzz
unidecode